C15 (v0.3.12): Refactored large files support
