+++ src/app/components/files-table/files-table.component.html
@@ -26,7 +26,7 @@
         </td>
         <td>{{cfile.path}}</td>
         <td>&nbsp;&nbsp;{{cfile.name}}</td>
-        <td style="color: var(--second-color)">{{cfile.encrypted ? "Encrypted" : ""}}</td>
+        <td style="color: var(--second-color)">&nbsp;&nbsp;{{cfile.encrypted ? "Encrypted" : ""}}</td>
         <td>{{cfile.formattedSize}}</td>
         <td class="buttons-cell">
           <div style="width: 90px">
